services/ai/requirements.txt
fastapi
uvicorn
pydantic
pika
google-genai
boto3